server/requirements.txt
fastapi==0.116.1
openai==1.108.1
psycopg[binary]==3.2.10
python-dotenv==1.1.1
uvicorn==0.35.0
pytest==8.4.2
